+++ node/clients/storefrontPermissions.ts
@@ -25,6 +25,24 @@ export const QUERIES = {
       canImpersonate
     }
   }`,
+  listUsersPaginated: `query users($organizationId: ID, $costCenterId: ID, $roleId: ID, $page: Int, $pageSize: Int, $search: String, $sortOrder: String, $sortedBy: String) {
+    listUsersPaginated(organizationId: $organizationId, costCenterId: $costCenterId, roleId: $roleId, page: $page, pageSize: $pageSize, search: $search, sortOrder: $sortOrder, sortedBy: $sortedBy) {
+      data {
+        id
+        roleId
+        userId
+        clId
+        orgId
+        costId
+        name
+        email
+        canImpersonate
+      }
+      pagination {
+        total
+      }
+    }
+  }`,
   listRoles: `query roles {
     listRoles {
       id
@@ -171,6 +189,49 @@ export default class StorefrontPermissions extends AppGraphQLClient {
         query: QUERIES.listUsers,
         variables: {
           roleId,
+          ...(organizationId && { organizationId }),
+          ...(costCenterId && { costCenterId }),
+        },
+        extensions: {
+          persistedQuery: {
+            provider: 'vtex.storefront-permissions@1.x',
+            sender: 'vtex.b2b-organizations@0.x',
+          },
+        },
+      },
+      {}
+    )
+  }
+
+  public listUsersPaginated = async ({
+    roleId,
+    organizationId,
+    costCenterId,
+    search = '',
+    page = 1,
+    pageSize = 25,
+    sortOrder = 'asc',
+    sortedBy = 'email',
+  }: {
+    roleId?: string
+    organizationId?: string
+    costCenterId?: string
+    search?: string
+    page?: number
+    pageSize?: number
+    sortOrder?: string
+    sortedBy?: string
+  }): Promise<any> => {
+    return this.graphql.query(
+      {
+        query: QUERIES.listUsersPaginated,
+        variables: {
+          roleId,
+          page,
+          pageSize,
+          search,
+          sortOrder,
+          sortedBy,
           ...(organizationId && { organizationId }),
           ...(costCenterId && { costCenterId }),
         },