+++ node/resolvers/Queries/index.ts
@@ -63,7 +63,6 @@ const Index = {
 
     return settings
   },
-
   getUsers: async (
     _: void,
     {
@@ -108,6 +107,103 @@ const Index = {
       .listUsers(variables)
       .then((result: any) => {
         return result.data.listUsers
+      })
+      .catch(error => {
+        logger.error({
+          message: 'getUsers-error',
+          error,
+        })
+        if (error.message) {
+          throw new GraphQLError(error.message)
+        } else if (error.response?.data?.message) {
+          throw new GraphQLError(error.response.data.message)
+        } else {
+          throw new GraphQLError(error)
+        }
+      })
+  },
+
+  getUsersPaginated: async (
+    _: void,
+    {
+      organizationId,
+      costCenterId,
+      search = '',
+      page = 1,
+      pageSize = 25,
+      sortOrder = 'asc',
+      sortedBy = 'email',
+    }: {
+      organizationId: string
+      costCenterId: string
+      search: string
+      page: number
+      pageSize: number
+      sortOrder: string
+      sortedBy: string
+    },
+    ctx: Context
+  ) => {
+    const {
+      clients: { storefrontPermissions },
+      vtex: { adminUserAuthToken, logger },
+      vtex,
+    } = ctx
+
+    const { sessionData } = vtex as any
+
+    const {
+      data: { checkUserPermission },
+    }: any = await storefrontPermissions
+      .checkUserPermission('vtex.b2b-organizations@1.x')
+      .catch((error: any) => {
+        logger.error({
+          error,
+          message: 'checkUserPermission-error',
+        })
+
+        return {
+          data: {
+            checkUserPermission: null,
+          },
+        }
+      })
+
+    const isSalesAdmin = checkUserPermission?.role.slug.match(/sales-admin/)
+
+    if (!adminUserAuthToken && !isSalesAdmin) {
+      if (!sessionData?.namespaces['storefront-permissions']) {
+        throw new GraphQLError('organization-data-not-found')
+      }
+
+      const {
+        organization: { value: userOrganizationId },
+      } = sessionData?.namespaces['storefront-permissions']
+
+      if (!organizationId) {
+        // get user's organization from session
+        organizationId = userOrganizationId
+      }
+
+      if (organizationId !== userOrganizationId) {
+        throw new GraphQLError('operation-not-permitted')
+      }
+    }
+
+    const variables = {
+      ...(organizationId && { organizationId }),
+      ...(costCenterId && { costCenterId }),
+      page,
+      pageSize,
+      search,
+      sortOrder,
+      sortedBy,
+    }
+
+    return storefrontPermissions
+      .listUsersPaginated(variables)
+      .then((result: any) => {
+        return result.data.listUsersPaginated
       })
       .catch(error => {
         logger.error({